refactor(locale): extract submodules from mod.rs
extract three submodules from the 2410-line locale/mod.rs:

- embedded/en_us.rs holds the hardcoded en-US baseline (role
  terms, locator terms, archive messages, vocab loader) that
  every other locale inherits from before applying overrides.
- raw_conversion.rs holds from_yaml_str / load / from_file plus
  the from_raw RawLocale → Locale transform and its parse/extract
  helpers, all on impl Locale.
- message_ids.rs holds the GeneralTerm / role / locator → message
  ID maps and the resolve_message_text MF2 dispatch.

mod.rs is now 1347 lines (-44%) and reads as the Locale struct +
its public term-lookup API. no behavior change; 1306/1306 tests
and clippy --workspace -D warnings pass.

closes item #4 on csl26-rfct.
