+++ crates/citum-schema-style/src/locale/mod.rs
@@ -8,12 +8,15 @@ SPDX-FileCopyrightText: © 2023-2026 Bruce D'Arcus
 //! Locales provide language-specific terms, date formats, and punctuation rules
 //! for citation formatting.
 
+mod embedded;
 /// Locator text normalization.
 pub mod locator;
 /// Message evaluation for parameterized locale strings.
 pub mod message;
+mod message_ids;
 /// Raw locale types used during locale file parsing.
 pub mod raw;
+mod raw_conversion;
 /// Structured locale types used by the processor.
 pub mod types;
 
@@ -26,7 +29,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fmt;
-use std::sync::{Arc, OnceLock};
+use std::sync::Arc;
 pub use types::*;
 
 /// A list of month names (12 elements for Jan-Dec).
@@ -129,87 +132,6 @@ fn default_evaluator() -> Arc<dyn MessageEvaluator> {
     Arc::new(Mf2MessageEvaluator)
 }
 
-#[derive(Deserialize)]
-struct EmbeddedVocabDocument {
-    #[serde(default)]
-    vocab: Option<raw::RawVocab>,
-}
-
-/// Extract one top-level YAML section while preserving its nested indentation.
-fn extract_top_level_yaml_section(yaml: &str, key: &str) -> Option<String> {
-    let header = format!("{key}:");
-    let mut collected = Vec::new();
-    let mut in_section = false;
-
-    for line in yaml.lines() {
-        let trimmed = line.trim_end_matches('\r');
-        let is_top_level =
-            !trimmed.is_empty() && !trimmed.starts_with(' ') && !trimmed.starts_with('\t');
-
-        if in_section {
-            if is_top_level {
-                break;
-            }
-            collected.push(trimmed);
-            continue;
-        }
-
-        if trimmed == header {
-            in_section = true;
-            collected.push(trimmed);
-        }
-    }
-
-    if collected.is_empty() {
-        None
-    } else {
-        Some(collected.join("\n"))
-    }
-}
-
-/// Archive hierarchy label messages for the hardcoded en-US locale.
-///
-/// Only the archive terms are pre-seeded here; all other message lookups fall
-/// through to the legacy typed term maps so that the hardcoded `en_us()`
-/// constructor stays consistent with the pre-existing test baseline.
-fn en_us_archive_messages() -> HashMap<String, String> {
-    [
-        ("term.archive-collection-label".into(), "collection".into()),
-        ("term.archive-series-label".into(), "series".into()),
-        (
-            "term.archive-box-label".into(),
-            ".match {$count :plural}\nwhen one {box}\nwhen * {boxes}".into(),
-        ),
-        (
-            "term.archive-folder-label".into(),
-            ".match {$count :plural}\nwhen one {folder}\nwhen * {folders}".into(),
-        ),
-        (
-            "term.archive-item-label".into(),
-            ".match {$count :plural}\nwhen one {item}\nwhen * {items}".into(),
-        ),
-    ]
-    .into()
-}
-
-/// Curated en-US genre and medium labels from the embedded locale asset.
-fn embedded_en_us_vocab() -> &'static VocabMap {
-    static EN_US_VOCAB: OnceLock<VocabMap> = OnceLock::new();
-
-    EN_US_VOCAB.get_or_init(|| {
-        crate::embedded::get_locale_bytes("en-US")
-            .and_then(|bytes| std::str::from_utf8(bytes).ok())
-            .and_then(|yaml| extract_top_level_yaml_section(yaml, "vocab"))
-            .and_then(|vocab_yaml| serde_yaml::from_str::<EmbeddedVocabDocument>(&vocab_yaml).ok())
-            .and_then(|document| document.vocab)
-            .map(|document| VocabMap {
-                genre: document.genre,
-                medium: document.medium,
-            })
-            .unwrap_or_default()
-    })
-}
-
 impl Default for Locale {
     fn default() -> Self {
         Self {
@@ -256,190 +178,6 @@ impl fmt::Debug for Locale {
     }
 }
 
-/// Extract English (US) role terms.
-fn en_us_role_terms() -> HashMap<ContributorRole, ContributorTerm> {
-    let mut roles = HashMap::new();
-
-    roles.insert(
-        ContributorRole::Editor,
-        ContributorTerm {
-            singular: SimpleTerm {
-                long: "editor".into(),
-                short: "ed.".into(),
-            },
-            plural: SimpleTerm {
-                long: "editors".into(),
-                short: "eds.".into(),
-            },
-            verb: SimpleTerm {
-                long: "edited by".into(),
-                short: "ed.".into(),
-            },
-        },
-    );
-
-    roles.insert(
-        ContributorRole::Translator,
-        ContributorTerm {
-            singular: SimpleTerm {
-                long: "translator".into(),
-                short: "Trans.".into(),
-            },
-            plural: SimpleTerm {
-                long: "translators".into(),
-                short: "Trans.".into(),
-            },
-            verb: SimpleTerm {
-                long: "translated by".into(),
-                short: "Trans.".into(),
-            },
-        },
-    );
-
-    roles.insert(
-        ContributorRole::Director,
-        ContributorTerm {
-            singular: SimpleTerm {
-                long: "director".into(),
-                short: "Dir.".into(),
-            },
-            plural: SimpleTerm {
-                long: "directors".into(),
-                short: "dirs.".into(),
-            },
-            verb: SimpleTerm {
-                long: "directed by".into(),
-                short: "dir.".into(),
-            },
-        },
-    );
-
-    roles.insert(
-        ContributorRole::Interviewer,
-        ContributorTerm {
-            singular: SimpleTerm {
-                long: "Interviewer".into(),
-                short: "Interviewer".into(),
-            },
-            plural: SimpleTerm {
-                long: "Interviewers".into(),
-                short: "Interviewers".into(),
-            },
-            verb: SimpleTerm {
-                long: "interviewed by".into(),
-                short: "interviewed by".into(),
-            },
-        },
-    );
-
-    roles
-}
-
-/// Extract English (US) locator terms.
-fn en_us_locator_terms() -> HashMap<LocatorType, LocatorTerm> {
-    let mut locators = HashMap::new();
-    locators.insert(
-        LocatorType::Page,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "page".into(),
-                plural: "pages".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "p.".into(),
-                plural: "pp.".into(),
-            }),
-            symbol: None,
-            gender: None,
-        },
-    );
-
-    locators.insert(
-        LocatorType::Chapter,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "chapter".into(),
-                plural: "chapters".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "ch.".into(),
-                plural: "chs.".into(),
-            }),
-            symbol: None,
-            gender: None,
-        },
-    );
-
-    locators.insert(
-        LocatorType::Volume,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "volume".into(),
-                plural: "volumes".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "vol.".into(),
-                plural: "vols.".into(),
-            }),
-            symbol: None,
-            gender: None,
-        },
-    );
-
-    locators.insert(
-        LocatorType::Section,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "section".into(),
-                plural: "sections".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "sec.".into(),
-                plural: "secs.".into(),
-            }),
-            symbol: Some(SingularPlural {
-                singular: "§".into(),
-                plural: "§§".into(),
-            }),
-            gender: None,
-        },
-    );
-
-    locators.insert(
-        LocatorType::Part,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "part".into(),
-                plural: "parts".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "pt.".into(),
-                plural: "pts.".into(),
-            }),
-            symbol: None,
-            gender: None,
-        },
-    );
-
-    locators.insert(
-        LocatorType::Supplement,
-        LocatorTerm {
-            long: Some(SingularPlural {
-                singular: "supplement".into(),
-                plural: "supplements".into(),
-            }),
-            short: Some(SingularPlural {
-                singular: "suppl.".into(),
-                plural: "suppls.".into(),
-            }),
-            symbol: None,
-            gender: None,
-        },
-    );
-
-    locators
-}
-
 /// Convert a kebab-case key to a human-readable display string.
 ///
 /// Splits on `-`, capitalizes the first character of the first word, and joins with spaces.
@@ -466,8 +204,8 @@ impl Locale {
         Self {
             locale: "en-US".into(),
             dates: DateTerms::en_us(),
-            roles: en_us_role_terms(),
-            locators: en_us_locator_terms(),
+            roles: embedded::en_us_role_terms(),
+            locators: embedded::en_us_locator_terms(),
             terms: Terms::en_us(),
             punctuation_in_quote: true,
             sort_articles: vec!["the".into(), "a".into(), "an".into()],
@@ -475,7 +213,7 @@ impl Locale {
             evaluation: EvaluationConfig {
                 message_syntax: MessageSyntax::Mf2,
             },
-            messages: en_us_archive_messages(),
+            messages: embedded::en_us_archive_messages(),
             date_formats: HashMap::new(),
             number_formats: NumberFormats {
                 decimal_separator: ".".into(),
@@ -493,7 +231,7 @@ impl Locale {
                 page_range_delimiter: "\u{2013}".into(),
             },
             legacy_term_aliases: HashMap::new(),
-            vocab: embedded_en_us_vocab().clone(),
+            vocab: embedded::embedded_en_us_vocab().clone(),
             evaluator: Arc::new(Mf2MessageEvaluator),
         }
     }
@@ -557,59 +295,6 @@ impl Locale {
             .get(key)
             .cloned()
             .unwrap_or_else(|| kebab_to_display(key))
-    }
-
-    /// Get default articles for a locale based on language code.
-    fn default_articles_for_locale(locale_id: &str) -> Vec<String> {
-        // Extract language code (first 2 chars)
-        #[allow(clippy::string_slice, reason = "locale_id is expected to be ASCII")]
-        let lang = &locale_id[..2.min(locale_id.len())];
-        match lang {
-            "en" => vec!["the".into(), "a".into(), "an".into()],
-            "de" => vec![
-                "der".into(),
-                "die".into(),
-                "das".into(),
-                "ein".into(),
-                "eine".into(),
-            ],
-            "fr" => vec![
-                "le".into(),
-                "la".into(),
-                "les".into(),
-                "l'".into(),
-                "un".into(),
-                "une".into(),
-            ],
-            "es" => vec![
-                "el".into(),
-                "la".into(),
-                "los".into(),
-                "las".into(),
-                "un".into(),
-                "una".into(),
-            ],
-            "it" => vec![
-                "il".into(),
-                "lo".into(),
-                "la".into(),
-                "i".into(),
-                "gli".into(),
-                "le".into(),
-                "un".into(),
-                "una".into(),
-            ],
-            "pt" => vec![
-                "o".into(),
-                "a".into(),
-                "os".into(),
-                "as".into(),
-                "um".into(),
-                "uma".into(),
-            ],
-            "nl" => vec!["de".into(), "het".into(), "een".into()],
-            _ => vec![], // Fall back to English defaults in strip_sort_articles
-        }
     }
 
     /// Get a contributor role term.
@@ -1004,168 +689,6 @@ impl Locale {
         }
     }
 
-    /// Map a GeneralTerm to its canonical message ID suffix (e.g., GeneralTerm::EtAl → "et-al").
-    fn general_term_to_message_id(term: &GeneralTerm) -> &str {
-        match term {
-            GeneralTerm::And => "and",
-            GeneralTerm::EtAl => "et-al",
-            GeneralTerm::AndOthers => "and-others",
-            GeneralTerm::Accessed => "accessed",
-            GeneralTerm::Retrieved => "retrieved",
-            GeneralTerm::NoDate => "no-date",
-            GeneralTerm::Ibid => "ibid",
-            GeneralTerm::In => "in",
-            GeneralTerm::At => "at",
-            GeneralTerm::By => "by",
-            GeneralTerm::From => "from",
-            GeneralTerm::Of => "of",
-            GeneralTerm::To => "to",
-            GeneralTerm::Anonymous => "anonymous",
-            GeneralTerm::Circa => "circa",
-            GeneralTerm::Forthcoming => "forthcoming",
-            GeneralTerm::Online => "online",
-            GeneralTerm::AvailableAt => "available-at",
-            GeneralTerm::ReviewOf => "review-of",
-            GeneralTerm::Here => "here",
-            GeneralTerm::Deposited => "deposited",
-            GeneralTerm::Patent => "patent",
-            GeneralTerm::Volume => "volume",
-            GeneralTerm::Issue => "issue",
-            GeneralTerm::Page => "page",
-            GeneralTerm::Chapter => "chapter",
-            GeneralTerm::Edition => "edition",
-            GeneralTerm::Section => "section",
-            GeneralTerm::OriginalWorkPublished => "original-work-published",
-            GeneralTerm::PersonalCommunication => "personal-communication",
-            GeneralTerm::Unknown(s) => s.as_str(),
-        }
-    }
-
-    /// Map a GeneralTerm to its legacy CSL key string for alias lookup.
-    fn general_term_to_legacy_key(term: &GeneralTerm) -> &str {
-        match term {
-            GeneralTerm::EtAl => "et_al",
-            GeneralTerm::NoDate => "no_date",
-            _ => Self::general_term_to_message_id(term),
-        }
-    }
-
-    fn role_message_id(role: &ContributorRole, form: &TermForm) -> Option<&'static str> {
-        let prefix = match role {
-            ContributorRole::Editor => "role.editor",
-            ContributorRole::Translator => "role.translator",
-            ContributorRole::Guest => "role.guest",
-            _ => return None,
-        };
-
-        match *form {
-            TermForm::Long => Some(match prefix {
-                "role.editor" => "role.editor.label-long",
-                "role.translator" => "role.translator.label-long",
-                "role.guest" => "role.guest.label-long",
-                _ => return None,
-            }),
-            TermForm::Short => Some(match prefix {
-                "role.editor" => "role.editor.label",
-                "role.translator" => "role.translator.label",
-                "role.guest" => "role.guest.label",
-                _ => return None,
-            }),
-            TermForm::Verb | TermForm::VerbShort => Some(match prefix {
-                "role.editor" => "role.editor.verb",
-                "role.translator" => "role.translator.verb",
-                "role.guest" => "role.guest.verb",
-                _ => return None,
-            }),
-            _ => None,
-        }
-    }
-
-    fn locator_message_id(locator: &LocatorType, form: &TermForm) -> Option<&'static str> {
-        let prefix = match locator {
-            LocatorType::Page => "term.page-label",
-            LocatorType::Chapter => "term.chapter-label",
-            LocatorType::Volume => "term.volume-label",
-            LocatorType::Section => "term.section-label",
-            LocatorType::Figure => "term.figure-label",
-            LocatorType::Note => "term.note-label",
-            _ => return None,
-        };
-
-        match *form {
-            TermForm::Long => Some(match prefix {
-                "term.page-label" => "term.page-label-long",
-                "term.chapter-label" => "term.chapter-label-long",
-                "term.volume-label" => "term.volume-label-long",
-                "term.section-label" => "term.section-label-long",
-                "term.figure-label" => "term.figure-label-long",
-                "term.note-label" => "term.note-label-long",
-                _ => return None,
-            }),
-            TermForm::Short => Some(prefix),
-            _ => None,
-        }
-    }
-
-    fn general_message_id(term: &GeneralTerm, form: &TermForm) -> Option<&'static str> {
-        match (term, form) {
-            (GeneralTerm::And, _) => Some("term.and"),
-            (GeneralTerm::EtAl, _) => Some("term.et-al"),
-            (GeneralTerm::AndOthers, _) => Some("term.and-others"),
-            (GeneralTerm::Accessed, _) => Some("term.accessed"),
-            (GeneralTerm::Retrieved, _) => Some("term.retrieved"),
-            (GeneralTerm::NoDate, TermForm::Long) => Some("term.no-date-long"),
-            (GeneralTerm::NoDate, _) => Some("term.no-date"),
-            (GeneralTerm::Forthcoming, _) => Some("term.forthcoming"),
-            (GeneralTerm::Circa, TermForm::Long) => Some("term.circa-long"),
-            (GeneralTerm::Circa, _) => Some("term.circa"),
-            _ => None,
-        }
-    }
-
-    fn gender_selector_key(gender: &GrammaticalGender) -> &str {
-        match gender {
-            GrammaticalGender::Masculine => "masculine",
-            GrammaticalGender::Feminine => "feminine",
-            GrammaticalGender::Neuter => "neuter",
-            GrammaticalGender::Common => "common",
-            GrammaticalGender::Unknown(s) => s.as_str(),
-        }
-    }
-
-    fn resolve_message_text(
-        &self,
-        message_id: &str,
-        count: Option<u64>,
-        gender: Option<GrammaticalGender>,
-    ) -> Option<String> {
-        let message = self.messages.get(message_id)?;
-
-        // Build MessageArgs for the evaluator
-        let args = MessageArgs {
-            count,
-            gender: gender.as_ref().map(Self::gender_selector_key),
-            ..MessageArgs::default()
-        };
-
-        // Store variables as owned Strings and then reference them
-        // (This is a limitation of the current design; could be improved with owned args)
-        // For now, we'll build a simple approach: just return the message if it's static
-        if !message.contains('{') {
-            return Some(message.clone());
-        }
-
-        // If no parameterized syntax, return None (fallback to legacy terms)
-        if self.evaluation.message_syntax == MessageSyntax::Static {
-            return None;
-        }
-
-        // Create a temporary variables map for the evaluator
-        // This is a simplification; the evaluator trait should ideally accept the args directly
-        // For now, we'll convert and handle in the message evaluator
-        self.evaluator.evaluate(message, &args)
-    }
-
     /// Resolve a `pattern.date-*` message with locale-specific year/month/day
     /// components.
     ///
@@ -1201,592 +724,6 @@ impl Locale {
             ..MessageArgs::default()
         };
         self.evaluator.evaluate(message, &args)
-    }
-}
-
-impl Locale {
-    /// Load a locale from a YAML string.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when the YAML cannot be parsed into a locale.
-    pub fn from_yaml_str(yaml: &str) -> Result<Self, String> {
-        let raw: raw::RawLocale = serde_yaml::from_str(yaml)
-            .map_err(|e| format!("Failed to parse locale YAML: {}", e))?;
-
-        Ok(Self::from_raw(raw))
-    }
-
-    /// Load a locale by ID (e.g., "en-US", "de-DE") from a locales directory.
-    /// Falls back to en-US if the locale file is not found.
-    pub fn load(locale_id: &str, locales_dir: &std::path::Path) -> Self {
-        let extensions = ["yaml", "yml", "json", "cbor"];
-
-        for ext in &extensions {
-            let file_name = format!("{}.{}", locale_id, ext);
-            let file_path = locales_dir.join(&file_name);
-
-            if file_path.exists() {
-                match Self::from_file(&file_path) {
-                    Ok(locale) => return locale,
-                    Err(e) => {
-                        eprintln!(
-                            "Warning: Failed to load locale {}.{}: {}",
-                            locale_id, ext, e
-                        );
-                    }
-                }
-            }
-        }
-
-        // Try fallback to base locale (e.g., "de" from "de-DE")
-        if locale_id.contains('-') {
-            let base = locale_id.split('-').next().unwrap_or("en");
-            // Try all files that start with base
-            if let Ok(entries) = std::fs::read_dir(locales_dir) {
-                for entry in entries.flatten() {
-                    let name = entry.file_name();
-                    let name_str = name.to_string_lossy();
-                    if (name_str.starts_with(base)
-                        && extensions.iter().any(|ext| name_str.ends_with(ext)))
-                        && let Ok(locale) = Self::from_file(&entry.path())
-                    {
-                        return locale;
-                    }
-                }
-            }
-        }
-
-        // Default to hardcoded en-US
-        Self::en_us()
-    }
-
-    /// Load locale from a file path directly (detects format).
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when the file cannot be read or its contents cannot be
-    /// parsed as a supported locale format.
-    pub fn from_file(path: &std::path::Path) -> Result<Self, String> {
-        let bytes =
-            std::fs::read(path).map_err(|e| format!("Failed to read locale file: {}", e))?;
-        let ext = path.extension().and_then(|e| e.to_str()).unwrap_or("yaml");
-
-        match ext {
-            "cbor" => ciborium::de::from_reader::<raw::RawLocale, _>(std::io::Cursor::new(&bytes))
-                .map(Self::from_raw)
-                .map_err(|e| format!("Failed to parse CBOR locale: {}", e)),
-            "json" => serde_json::from_slice::<raw::RawLocale>(&bytes)
-                .map(Self::from_raw)
-                .map_err(|e| format!("Failed to parse JSON locale: {}", e)),
-            _ => {
-                let content = String::from_utf8_lossy(&bytes);
-                Self::from_yaml_str(&content)
-            }
-        }
-    }
-
-    /// Convert a RawLocale to a Locale.
-    #[allow(
-        clippy::too_many_lines,
-        reason = "Complex parsing of raw locale data with multiple term types"
-    )]
-    fn from_raw(raw: raw::RawLocale) -> Self {
-        // Determine punctuation-in-quote from locale ID
-        // en-US uses American style (inside), en-GB and others use outside
-        let punctuation_in_quote = raw.locale.starts_with("en-US")
-            || (raw.locale.starts_with("en") && !raw.locale.starts_with("en-GB"));
-
-        // Start from en-US defaults so partially specified locale files still
-        // have complete term/locator coverage (e.g., page/section labels).
-        let mut locale = Locale::en_us();
-        locale.locale = raw.locale.clone();
-        locale.dates = DateTerms {
-            months: MonthNames {
-                long: raw.dates.months.long,
-                short: raw.dates.months.short,
-            },
-            seasons: raw.dates.seasons,
-            uncertainty_term: raw.dates.uncertainty_term,
-            open_ended_term: raw.dates.open_ended_term,
-            am: raw.dates.am,
-            pm: raw.dates.pm,
-            timezone_utc: raw.dates.timezone_utc,
-            before_era: raw.dates.before_era,
-            ad: raw.dates.ad,
-            bc: raw.dates.bc,
-            bce: raw.dates.bce,
-            ce: raw.dates.ce,
-        };
-        locale.punctuation_in_quote = punctuation_in_quote;
-        // Set locale-specific articles based on language
-        locale.sort_articles = Self::default_articles_for_locale(&raw.locale);
-
-        // v2 schema fields — copied verbatim (no structural transformation needed at this layer)
-        locale.locale_schema_version = raw.locale_schema_version;
-        locale.evaluation = raw.evaluation.unwrap_or_default();
-        locale.messages = raw.messages;
-        locale.date_formats = raw.date_formats;
-        locale.legacy_term_aliases = raw.legacy_term_aliases;
-
-        // Merge vocab overrides into the embedded en-US defaults.
-        if let Some(raw_vocab) = raw.vocab {
-            locale.vocab.genre.extend(raw_vocab.genre);
-            locale.vocab.medium.extend(raw_vocab.medium);
-        }
-
-        // Merge grammar_options: use raw value if present, otherwise derive from locale ID
-        if let Some(go) = raw.grammar_options {
-            locale.grammar_options = go;
-        } else {
-            // Derive punctuation_in_quote from locale ID (preserving existing behaviour)
-            locale.grammar_options.punctuation_in_quote = locale.punctuation_in_quote;
-        }
-        // For v2 locales that explicitly declare grammar_options, the grammar_options
-        // field is the authoritative source. Sync back to the legacy punctuation_in_quote
-        // field so all existing call sites in citum-engine get the correct value.
-        locale.punctuation_in_quote = locale.grammar_options.punctuation_in_quote;
-
-        // Merge number_formats if provided
-        if let Some(nf) = raw.number_formats {
-            locale.number_formats = nf;
-        }
-
-        let explicit_locator_keys: std::collections::HashSet<LocatorType> = raw
-            .locators
-            .keys()
-            .filter_map(|key| Self::parse_builtin_locator_type(key))
-            .collect();
-
-        for (key, value) in &raw.locators {
-            if let Some(locator_type) = Self::parse_locator_type(key) {
-                let locator_term = LocatorTerm {
-                    long: Self::extract_singular_plural(value.long.as_ref().as_ref()),
-                    short: Self::extract_singular_plural(value.short.as_ref().as_ref()),
-                    symbol: Self::extract_singular_plural(value.symbol.as_ref().as_ref()),
-                    gender: value.gender.clone(),
-                };
-                locale.locators.insert(locator_type, locator_term);
-            }
-        }
-
-        // Map raw terms to structured general terms.
-        for (key, value) in &raw.terms {
-            if let Some(locator_type) = Self::parse_builtin_locator_type(key)
-                && !explicit_locator_keys.contains(&locator_type)
-                && let Some(forms) = Self::get_forms(value)
-            {
-                let locator_term = LocatorTerm {
-                    long: Self::extract_singular_plural(forms.get("long").as_ref()),
-                    short: Self::extract_singular_plural(forms.get("short").as_ref()),
-                    symbol: Self::extract_singular_plural(forms.get("symbol").as_ref()),
-                    gender: None,
-                };
-                locale.locators.insert(locator_type, locator_term);
-                continue;
-            }
-
-            match key.as_str() {
-                "and" => {
-                    if let Some(forms) = Self::get_forms(value) {
-                        if let Some(v) = forms.get("long").and_then(|v| v.as_string()) {
-                            locale.terms.and = Some(v.to_string());
-                        }
-                        if let Some(v) = forms.get("symbol").and_then(|v| v.as_string()) {
-                            locale.terms.and_symbol = Some(v.to_string());
-                        }
-                    }
-                }
-                "et_al" => {
-                    if let Some(forms) = Self::get_forms(value)
-                        && let Some(v) = forms.get("long").and_then(|v| v.as_string())
-                    {
-                        locale.terms.et_al = Some(v.to_string());
-                    }
-                }
-                "and others" | "and_others" => {
-                    if let Some(forms) = Self::get_forms(value)
-                        && let Some(v) = forms.get("long").and_then(|v| v.as_string())
-                    {
-                        locale.terms.and_others = Some(v.to_string());
-                    }
-                }
-                "accessed" => {
-                    if let Some(forms) = Self::get_forms(value)
-                        && let Some(v) = forms.get("long").and_then(|v| v.as_string())
-                    {
-                        locale.terms.accessed = Some(v.to_string());
-                    }
-                }
-                "ibid" => {
-                    if let Some(forms) = Self::get_forms(value)
-                        && let Some(v) = forms.get("long").and_then(|v| v.as_string())
-                    {
-                        locale.terms.ibid = Some(v.to_string());
-                    }
-                }
-                "no date" => {
-                    let simple = Self::extract_simple_term_from_raw(value);
-                    let short_fallback = simple.short.as_default_str().to_string();
-                    locale.terms.general.insert(GeneralTerm::NoDate, simple);
-                    locale.terms.no_date.get_or_insert(short_fallback);
-                }
-                "no_date" => {
-                    let simple = Self::extract_simple_term_from_raw(value);
-                    locale.terms.no_date = Some(simple.short.as_str().to_string());
-                    locale
-                        .terms
-                        .general
-                        .entry(GeneralTerm::NoDate)
-                        .or_insert(simple);
-                }
-                _ => {
-                    // Try to parse as GeneralTerm
-                    if let Some(general_term) = Self::parse_general_term(key) {
-                        let simple = Self::extract_simple_term_from_raw(value);
-                        locale.terms.general.insert(general_term, simple);
-                    }
-                }
-            }
-        }
-
-        // Map raw roles to structured roles (simplified for now)
-        for (key, role_term) in &raw.roles {
-            if let Some(role) = Self::parse_role_name(key) {
-                let contributor_term = ContributorTerm {
-                    singular: Self::extract_simple_term(&role_term.long, &role_term.short, false),
-                    plural: Self::extract_simple_term(&role_term.long, &role_term.short, true),
-                    verb: Self::extract_verb_term(&role_term.verb, &role_term.verb_short),
-                };
-                locale.roles.insert(role, contributor_term);
-            }
-        }
-
-        // Set the message evaluator based on evaluation.message_syntax
-        locale.evaluator = match locale.evaluation.message_syntax {
-            MessageSyntax::Mf2 => Arc::new(Mf2MessageEvaluator),
-            MessageSyntax::Static => Arc::new(Mf2MessageEvaluator),
-        };
-
-        locale
-    }
-
-    fn get_forms(value: &raw::RawTermValue) -> Option<&HashMap<String, raw::RawTermValue>> {
-        match value {
-            raw::RawTermValue::Forms(forms) => Some(forms),
-            _ => None,
-        }
-    }
-
-    fn parse_locator_type(name: &str) -> Option<LocatorType> {
-        LocatorType::from_key(name).ok()
-    }
-
-    fn parse_builtin_locator_type(name: &str) -> Option<LocatorType> {
-        match Self::parse_locator_type(name)? {
-            LocatorType::Custom(_) => None,
-            locator => Some(locator),
-        }
-    }
-
-    fn parse_role_name(name: &str) -> Option<ContributorRole> {
-        match name {
-            "author" => Some(ContributorRole::Author),
-            "chair" => Some(ContributorRole::Chair),
-            "editor" => Some(ContributorRole::Editor),
-            "translator" => Some(ContributorRole::Translator),
-            "director" => Some(ContributorRole::Director),
-            "compiler" => Some(ContributorRole::Composer), // Close mapping
-            "illustrator" => Some(ContributorRole::Illustrator),
-            "collection-editor" => Some(ContributorRole::CollectionEditor),
-            "container-author" => Some(ContributorRole::ContainerAuthor),
-            "editorial-director" => Some(ContributorRole::EditorialDirector),
-            "textual-editor" | "textual_editor" => Some(ContributorRole::TextualEditor),
-            "interviewer" => Some(ContributorRole::Interviewer),
-            "original-author" => Some(ContributorRole::OriginalAuthor),
-            "recipient" => Some(ContributorRole::Recipient),
-            "reviewed-author" => Some(ContributorRole::ReviewedAuthor),
-            "composer" => Some(ContributorRole::Composer),
-            _ => None,
-        }
-    }
-
-    fn extract_singular_plural(value: Option<&&raw::RawTermValue>) -> Option<SingularPlural> {
-        match value {
-            Some(raw::RawTermValue::SingularPlural { singular, plural }) => Some(SingularPlural {
-                singular: Self::from_raw_gendered_string(singular),
-                plural: Self::from_raw_gendered_string(plural),
-            }),
-            Some(raw::RawTermValue::Simple(s)) => Some(SingularPlural {
-                singular: MaybeGendered::Plain(s.clone()),
-                plural: MaybeGendered::Plain(s.clone()), // Fallback if only one form provided
-            }),
-            Some(raw::RawTermValue::Gendered {
-                masculine,
-                feminine,
-                neuter,
-                common,
-            }) => Some(SingularPlural {
-                singular: MaybeGendered::Gendered {
-                    masculine: masculine.clone(),
-                    feminine: feminine.clone(),
-                    neuter: neuter.clone(),
-                    common: common.clone(),
-                },
-                plural: MaybeGendered::Gendered {
-                    masculine: masculine.clone(),
-                    feminine: feminine.clone(),
-                    neuter: neuter.clone(),
-                    common: common.clone(),
-                },
-            }),
-            Some(raw::RawTermValue::Forms(forms)) => {
-                let singular = forms
-                    .get("singular")
-                    .map(Self::extract_maybe_gendered_string);
-                let plural = forms.get("plural").map(Self::extract_maybe_gendered_string);
-
-                singular.map(|s| SingularPlural {
-                    plural: plural.unwrap_or_else(|| s.clone()),
-                    singular: s,
-                })
-            }
-            _ => None,
-        }
-    }
-
-    fn extract_simple_term(
-        long: &Option<raw::RawTermValue>,
-        short: &Option<raw::RawTermValue>,
-        plural: bool,
-    ) -> SimpleTerm {
-        let long_str = long
-            .as_ref()
-            .map(|v| Self::extract_simple_gendered_term(v, plural))
-            .unwrap_or_default();
-
-        let short_str = short
-            .as_ref()
-            .map(|v| Self::extract_simple_gendered_term(v, plural))
-            .unwrap_or_default();
-
-        SimpleTerm {
-            long: long_str,
-            short: short_str,
-        }
-    }
-
-    fn extract_verb_term(
-        verb: &Option<raw::RawTermValue>,
-        verb_short: &Option<raw::RawTermValue>,
-    ) -> SimpleTerm {
-        let long_str = verb
-            .as_ref()
-            .and_then(|v| v.as_string())
-            .unwrap_or("")
-            .into();
-
-        let short_str = verb_short
-            .as_ref()
-            .and_then(|v| v.as_string())
-            .unwrap_or("")
-            .into();
-
-        SimpleTerm {
-            long: long_str,
-            short: short_str,
-        }
-    }
-
-    /// Normalize a locale term key to canonical kebab-case.
-    ///
-    /// Locale YAML files and style templates may use underscores or spaces
-    /// interchangeably with hyphens (e.g. `no_date`, `no date`, `no-date`).
-    /// This helper converts all three forms to the single canonical
-    /// kebab-case key so `parse_general_term` only needs to match one pattern
-    /// per term.
-    fn normalize_term_key(s: &str) -> String {
-        s.replace(['_', ' '], "-")
-    }
-
-    /// Parse a locale term key into a structured general-term identifier.
-    pub fn parse_general_term(name: &str) -> Option<GeneralTerm> {
-        match Self::normalize_term_key(name).as_str() {
-            "in" => Some(GeneralTerm::In),
-            "accessed" => Some(GeneralTerm::Accessed),
-            "retrieved" => Some(GeneralTerm::Retrieved),
-            "at" => Some(GeneralTerm::At),
-            "from" => Some(GeneralTerm::From),
-            "of" => Some(GeneralTerm::Of),
-            "to" => Some(GeneralTerm::To),
-            "by" => Some(GeneralTerm::By),
-            "no-date" => Some(GeneralTerm::NoDate),
-            "anonymous" => Some(GeneralTerm::Anonymous),
-            "circa" => Some(GeneralTerm::Circa),
-            "available-at" => Some(GeneralTerm::AvailableAt),
-            "ibid" => Some(GeneralTerm::Ibid),
-            "and" => Some(GeneralTerm::And),
-            "et-al" => Some(GeneralTerm::EtAl),
-            "and-others" => Some(GeneralTerm::AndOthers),
-            "forthcoming" => Some(GeneralTerm::Forthcoming),
-            "online" => Some(GeneralTerm::Online),
-            "here" => Some(GeneralTerm::Here),
-            "deposited" => Some(GeneralTerm::Deposited),
-            "review-of" => Some(GeneralTerm::ReviewOf),
-            "original-work-published" => Some(GeneralTerm::OriginalWorkPublished),
-            "personal-communication" => Some(GeneralTerm::PersonalCommunication),
-            "patent" => Some(GeneralTerm::Patent),
-            "volume" => Some(GeneralTerm::Volume),
-            "issue" => Some(GeneralTerm::Issue),
-            "page" => Some(GeneralTerm::Page),
-            "chapter" => Some(GeneralTerm::Chapter),
-            "edition" => Some(GeneralTerm::Edition),
-            "section" => Some(GeneralTerm::Section),
-            _ => None,
-        }
-    }
-
-    fn extract_simple_term_from_raw(value: &raw::RawTermValue) -> SimpleTerm {
-        match value {
-            raw::RawTermValue::Simple(s) => SimpleTerm {
-                long: s.clone().into(),
-                short: s.clone().into(),
-            },
-            raw::RawTermValue::Gendered {
-                masculine,
-                feminine,
-                neuter,
-                common,
-            } => SimpleTerm {
-                long: MaybeGendered::Gendered {
-                    masculine: masculine.clone(),
-                    feminine: feminine.clone(),
-                    neuter: neuter.clone(),
-                    common: common.clone(),
-                },
-                short: MaybeGendered::Gendered {
-                    masculine: masculine.clone(),
-                    feminine: feminine.clone(),
-                    neuter: neuter.clone(),
-                    common: common.clone(),
-                },
-            },
-            raw::RawTermValue::Forms(forms) => {
-                let long = forms
-                    .get("long")
-                    .map(Self::extract_maybe_gendered_string)
-                    .unwrap_or_default();
-                let short = forms
-                    .get("short")
-                    .map(Self::extract_maybe_gendered_string)
-                    .unwrap_or_else(|| long.clone());
-                SimpleTerm { long, short }
-            }
-            raw::RawTermValue::SingularPlural { singular, .. } => SimpleTerm {
-                long: Self::from_raw_gendered_string(singular),
-                short: Self::from_raw_gendered_string(singular),
-            },
-        }
-    }
-
-    fn from_raw_gendered_string(value: &raw::RawGenderedString) -> MaybeGendered<String> {
-        match value {
-            raw::RawGenderedString::Simple(value) => MaybeGendered::Plain(value.clone()),
-            raw::RawGenderedString::Gendered {
-                masculine,
-                feminine,
-                neuter,
-                common,
-            } => MaybeGendered::Gendered {
-                masculine: masculine.clone(),
-                feminine: feminine.clone(),
-                neuter: neuter.clone(),
-                common: common.clone(),
-            },
-        }
-    }
-
-    fn extract_maybe_gendered_string(value: &raw::RawTermValue) -> MaybeGendered<String> {
-        match value {
-            raw::RawTermValue::Simple(value) => MaybeGendered::Plain(value.clone()),
-            raw::RawTermValue::Gendered {
-                masculine,
-                feminine,
-                neuter,
-                common,
-            } => MaybeGendered::Gendered {
-                masculine: masculine.clone(),
-                feminine: feminine.clone(),
-                neuter: neuter.clone(),
-                common: common.clone(),
-            },
-            raw::RawTermValue::SingularPlural { singular, .. } => {
-                Self::from_raw_gendered_string(singular)
-            }
-            raw::RawTermValue::Forms(forms) => forms
-                .get("long")
-                .or_else(|| forms.get("singular"))
-                .map(Self::extract_maybe_gendered_string)
-                .unwrap_or_default(),
-        }
-    }
-
-    fn extract_simple_gendered_term(
-        value: &raw::RawTermValue,
-        plural: bool,
-    ) -> MaybeGendered<String> {
-        match value {
-            raw::RawTermValue::Simple(value) => MaybeGendered::Plain(value.clone()),
-            raw::RawTermValue::Gendered {
-                masculine,
-                feminine,
-                neuter,
-                common,
-            } => MaybeGendered::Gendered {
-                masculine: masculine.clone(),
-                feminine: feminine.clone(),
-                neuter: neuter.clone(),
-                common: common.clone(),
-            },
-            raw::RawTermValue::SingularPlural {
-                singular,
-                plural: plural_value,
-            } => {
-                if plural {
-                    Self::from_raw_gendered_string(plural_value)
-                } else {
-                    Self::from_raw_gendered_string(singular)
-                }
-            }
-            raw::RawTermValue::Forms(forms) => {
-                let key = if plural { "plural" } else { "singular" };
-                forms
-                    .get(key)
-                    .or_else(|| forms.get("long"))
-                    .map(Self::extract_maybe_gendered_string)
-                    .unwrap_or_default()
-            }
-        }
-    }
-
-    /// Apply a partial override, merging its fields into this locale.
-    ///
-    /// Performs key-by-key insertion or replacement for:
-    /// - `messages`: new or updated message IDs
-    /// - `grammar_options`: if `Some`, replaces the entire block and syncs
-    ///   `punctuation_in_quote` field
-    /// - `legacy_term_aliases`: new or updated term aliases
-    pub fn apply_override(&mut self, ov: &LocaleOverride) {
-        for (k, v) in &ov.messages {
-            self.messages.insert(k.clone(), v.clone());
-        }
-        if let Some(go) = &ov.grammar_options {
-            self.grammar_options = go.clone();
-            self.punctuation_in_quote = go.punctuation_in_quote;
-        }
-        for (k, v) in &ov.legacy_term_aliases {
-            self.legacy_term_aliases.insert(k.clone(), v.clone());
-        }
     }
 }
